refactor: name modules after packages
