+++ crates/wasi/src/p2/host/network.rs
@@ -248,7 +248,7 @@ pub(crate) mod util {
     use std::net::{IpAddr, Ipv6Addr, SocketAddr};
     use std::time::Duration;
 
-    use crate::net::SocketAddressFamily;
+    use crate::sockets::SocketAddressFamily;
     use cap_net_ext::{AddressFamily, Blocking, UdpSocketExt};
     use rustix::fd::{AsFd, OwnedFd};
     use rustix::io::Errno;
