+++ crates/wasi/src/p2/tcp.rs
@@ -1,4 +1,3 @@
-use crate::net::{DEFAULT_TCP_BACKLOG, SocketAddressFamily};
 use crate::p2::bindings::sockets::tcp::ErrorCode;
 use crate::p2::host::network;
 use crate::p2::{
@@ -6,6 +5,7 @@ use crate::p2::{
     SocketResult, StreamError,
 };
 use crate::runtime::{AbortOnDropJoinHandle, with_ambient_tokio_runtime};
+use crate::sockets::{DEFAULT_TCP_BACKLOG, SocketAddressFamily};
 use anyhow::Result;
 use cap_net_ext::AddressFamily;
 use futures::Future;
